refactor: move pipline to different file
