+++ src/tools/commands/commands.js
@@ -1,26 +1,26 @@
-import { stdout } from "node:process";
-import { directory } from "../cwd/cwd.js";
+// import { stdout } from "node:process";
+// import { directory } from "../cwd/cwd.js";
 
-const tempCommand = () => {
-  stdout.write("\ndddd\n");
-};
+// const tempCommand = () => {
+//   stdout.write("\ndddd\n");
+// };
 
-const commands = {
-  ls: directory.ls,
-  cd: async (arg) => {
-    await directory.cd(arg);
-  },
-  up: directory.up,
-  cat: tempCommand,
-  add: tempCommand,
-  rn: tempCommand,
-  cp: tempCommand,
-  mv: tempCommand,
-  rm: tempCommand,
-  os: tempCommand,
-  hash: tempCommand,
-  compress: tempCommand,
-  decompress: tempCommand,
-};
+// const commands = {
+//   ls: directory.ls,
+//   cd: async (arg) => {
+//     await directory.cd(arg);
+//   },
+//   up: directory.up,
+//   cat: tempCommand,
+//   add: tempCommand,
+//   rn: tempCommand,
+//   cp: tempCommand,
+//   mv: tempCommand,
+//   rm: tempCommand,
+//   os: tempCommand,
+//   hash: tempCommand,
+//   compress: tempCommand,
+//   decompress: tempCommand,
+// };
 
-export { commands };
+// export { commands };
